# cmd/server/docs/swagger.yaml
basePath: /
definitions:
  dto.Avatar:
    properties:
      id:
        type: string
      image:
        type: string
      private:
        type: boolean
    type: object
...
  dto.EquipmentItem:
    properties:
      artifact:
        type: boolean
      attack:
        type: integer
      createdAt:
        type: string
      defense:
        type: integer
      hp:
        type: integer
      id:
        type: string
      image:
        type: string
      name:
        type: string
      price:
        type: integer
      requiredLevel:
        type: integer
      slug:
        type: string
    type: object
  dto.UpdateUserRequest:
    properties:
      avatarId:
        type: string
    type: object
  dto.User:
    properties:
      armsEquipmentItemId:
        type: string
      attack:
        type: integer
      avatar:
        $ref: '#/definitions/dto.Avatar'
      beltEquipmentItemId:
        type: string
      chestEquipmentItemId:
        type: string
      createdAt:
        type: string
      currentHp:
        type: integer
      defense:
        type: integer
      email:
        type: string
      exp:
        type: integer
      feetEquipmentItemId:
        type: string
      freeStats:
        type: integer
      gold:
        type: integer
      handsEquipmentItemId:
        type: string
      headEquipmentItemId:
        type: string
      hp:
        type: integer
      id:
        type: string
      inFight:
        type: boolean
      legsEquipmentItemId:
        type: string
      level:
        type: integer
      locationSlug:
        type: string
      neckEquipmentItemId:
        type: string
      ring1EquipmentItemId:
        type: string
      ring2EquipmentItemId:
        type: string
      ring3EquipmentItemId:
        type: string
      ring4EquipmentItemId:
        type: string
      shieldEquipmentItemId:
        type: string
      username:
        type: string
      weaponEquipmentItemId:
        type: string
    type: object
...
  handlers.LocationCellsResponse:
    properties:
      cells:
        items:
          $ref: '#/definitions/handlers.locationCell'
        type: array
    type: object
  handlers.MoveToCellResponse:
    properties:
      message:
        type: string
      path_length:
        type: integer
    type: object
...
  handlers.locationCell:
    properties:
      id:
        type: string
      image:
        type: string
      inactive:
        type: boolean
      name:
        type: string
      slug:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Game API for Moonshine
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Moonshine API
  version: "1.0"
paths:
...
  /api/locations/{slug}/cells:
    get:
      consumes:
      - application/json
      description: Get list of cells in a location
      parameters:
      - description: Location slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LocationCellsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get location cells
      tags:
      - locations
  /api/locations/{slug}/cells/{cell_slug}/move:
    post:
      consumes:
      - application/json
      description: Start movement to a cell within location
      parameters:
      - description: Location slug
        in: path
        name: slug
        required: true
        type: string
      - description: Cell slug
        in: path
        name: cell_slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MoveToCellResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Move to cell
      tags:
      - locations
  /api/locations/{slug}/move:
    post:
      consumes:
      - application/json
      description: Move user to a different location
      parameters:
      - description: Location slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Move to location
      tags:
      - locations
  /api/user/me:
    get:
      consumes:
      - application/json
      description: Get authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get current user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update authenticated user information
      parameters:
      - description: Update user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update current user
      tags:
      - user
  /api/users/me/equipped:
    get:
      consumes:
      - application/json
      description: Get list of currently equipped items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/dto.EquipmentItem'
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get equipped items
      tags:
      - user
  /api/users/me/inventory:
    get:
      consumes:
      - application/json
      description: Get list of items in user's inventory
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.EquipmentItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get user inventory
      tags:
      - user
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 'JWT token. Example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"